Massaging filenames to minimize diff.
